feat: score compose
